requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
discord.py==2.4.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
multidict==6.1.0
mysqlclient==2.2.5
propcache==0.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
yarl==1.15.2
